Do not delete the source image when finalizing docker CFPackages
In contrast to buildpack packages where Korifi creates the source image,
the source image for docker packages has been created by someone in
advance. Therefore Korifi should not try to clean it up on CFPackage
finalization.

+++ controllers/controllers/workloads/cfpackage_controller.go
@@ -148,7 +148,7 @@ func (r *CFPackageReconciler) finalize(ctx context.Context, cfPackage *korifiv1a
 		return ctrl.Result{}, nil
 	}
 
-	if cfPackage.Spec.Source.Registry.Image != "" {
+	if cfPackage.Spec.Type != "docker" && cfPackage.Spec.Source.Registry.Image != "" {
 		if err := r.imageDeleter.Delete(ctx, image.Creds{
 			Namespace:   cfPackage.Namespace,
 			SecretNames: r.packageRepoSecretNames,
